[ONEM-25211] UnicodeDecodeError fix.
Repo reads git output in chunks. It is possible that chunk
ends in the middle of uft-8 sequence, which leads to decod eerror.

Unicode decoding errors are now switched to replace mode.
Please see Python codecs documentation for details.
Not possible to decode characters are replaced with special '?'.

At the same time read chunk size has been increased to 16kB.